requirements.txt
streamlit>=1.40.0
pandas>=2.2.3
openpyxl>=3.1.5
python-barcode>=0.15.1
opencv-python-headless>=4.10.0.84
pyzbar>=0.1.9
pytest>=8.3.0
pytest-cov>=6.0.0
python-dotenv>=1.0.1
pillow>=11.0.0